Swap test files. Improve config change
